requirements.txt
blinker==1.7.0
cbor2==5.5.1
click==8.1.7
colorama==0.4.6
distlib==0.3.6
filelock==3.12.2
Flask==3.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
luma.core==2.4.1
luma.oled==3.13.0
MarkupSafe==2.1.3
Pillow==10.1.0
platformdirs==3.5.3
pyftdi==0.55.0
pyserial==3.5
pyusb==1.2.1
PyYAML==6.0.1
smbus2==0.4.3
virtualenv==20.23.0
Werkzeug==3.0.1
